Added files with updated code
"2nd competition code.txt" has the edits we made during the second competition, as well as removing the unused drive motors and adding a fast and slow speed.

The single player file is a modification of the original file that allow you to drive and codrive the robot on one controller.